backend/requirements.txt
# Web Framework ve API
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
email-validator==2.1.0.post1

# Veritabanı
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Kimlik Doğrulama ve Güvenlik
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==41.0.5

# AI Model ve GPU
torch==2.1.0
transformers==4.35.2
huggingface-hub==0.18.0
optimum==1.13.2
onnxruntime==1.16.1
onnxruntime-gpu==1.16.1
nvidia-ml-py==12.535.77

# Önbellek ve İzleme
redis==5.0.1
prometheus-client==0.18.0

# Vektör Veritabanı
qdrant-client==1.6.3

# Yardımcı Araçlar
python-dotenv==1.0.0
tenacity==8.2.3
httpx==0.25.1
numpy==1.26.1
pandas==2.1.2
Pillow==10.1.0
scikit-learn==1.3.2